requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn==20.1.0
django-cors-headers
django-bootstrap5
psycopg[binary]